+++ CHANGES.md
@@ -27,6 +27,7 @@
 * added setting to display `_deprecated_hook()` data in debug.log
 * skip `_deprecated_hook()` `trigger_error()` in development environment
 * zero value of repo cache release asset `$url` if `wp_remote_get( $url )` not HTTP code 200 when checking release asset redirect
+* add **Add-Ons** tab for installing API plugins
 
 #### 9.9.10 / 2021-02-18
 * fix change to `redirect_on_save()`